requirements.txt
openai>=1.0.0
anthropic>=0.3.0
python-dotenv>=0.19.0
jinja2>=3.0.0
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.65.0
langchain-community>=0.2.0
scipy>=1.10.0